Refactor user journey to be "Apply for ELM"
This seems more believable than the generic "Select a business" that was
used previously.
