+++ app/plugins/router.js
@@ -1,6 +1,6 @@
 const routes = [].concat(
-  require('../routes/business'),
-  require('../routes/business/view'),
+  require('../routes/apply/business'),
+  require('../routes/apply/select-business'),
   require('../routes/home'),
   require('../routes/healthy'),
   require('../routes/healthz'),
