+++ app/routes/apply/business.js
@@ -1,9 +1,9 @@
-const getBusinessViewModel = require('../../view-models/business/view')
+const getBusinessViewModel = require('../../view-models/apply/business')
 const landService = require('../../services/land-service')
 
 module.exports = {
   method: 'GET',
-  path: '/business/{sbi}',
+  path: '/apply/{sbi}',
   options: {
     handler: async (request, h) => {
       try {
@@ -11,7 +11,7 @@ module.exports = {
 
         const parcels = await landService.getParcels(sbi)
 
-        return h.view('business/view', getBusinessViewModel({ parcels, sbi }))
+        return h.view('apply/business', getBusinessViewModel({ parcels, sbi }))
       } catch (error) {
         console.error(error)
         return h.view('500')
