+++ app/routes/apply/select-business.js
@@ -1,13 +1,13 @@
 module.exports = {
   method: 'GET',
-  path: '/business',
+  path: '/apply',
   options: {
     handler: async (request, h) => {
       try {
         const businesses = [
           { sbi: 106599008 }
         ]
-        return h.view('business/list', { businesses })
+        return h.view('apply/select-business', { businesses })
       } catch (error) {
         console.error(error)
         return h.view('500')
