+++ app/views/apply/business.njk
@@ -13,12 +13,12 @@
         href: "/"
       },
       {
-        text: "Select business",
-        href: "/business"
+        text: "Apply for ELM",
+        href: "/apply"
       },
       {
         text: sbi,
-        href: "/business/" + sbi
+        href: "/apply/" + sbi
       }
     ]
   }) }}
