+++ app/views/apply/select-business.njk
@@ -13,8 +13,8 @@
         href: "/"
       },
       {
-        text: "Select business",
-        href: "/business"
+        text: "Apply for ELM",
+        href: "/apply"
       }
     ]
   }) }}
@@ -23,7 +23,7 @@
 {% block content %}
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
-      <h1 class="govuk-heading-xl">Select business</h1>
+      <h1 class="govuk-heading-xl">Apply for ELM</h1>
 
       <p class="govuk-body">
         Select from the list below which business you would like to apply for ELM on behalf of.
@@ -35,7 +35,7 @@
 
       <ul class="govuk-list">
       {% for business in businesses %}
-        <a href="/business/{{ business.sbi }}" class="govuk-link">{{ business.sbi }}</a>
+        <a href="/apply/{{ business.sbi }}" class="govuk-link">{{ business.sbi }}</a>
       {% endfor %}
       </ul>
     </div>
