+++ app/views/home.njk
@@ -9,7 +9,7 @@
     items: [
       {
         text: "Home",
-        href: "#"
+        href: "/"
       }]
   }) }}
 {% endblock %}
@@ -31,7 +31,7 @@
 
       {{ govukButton({
         classes: "govuk-!-margin-top-2 govuk-!-margin-bottom-8",
-        href: "/business",
+        href: "/apply",
         isStartButton: true,
         text: "Start now"
       }) }}
